Backend-django/requirements.txt
Django==3.1.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.8.6
djangorestframework-simplejwt==4.7.1
django-cors-headers==3.7.0